fix: time generator
