+++ src/global_vars.hpp
@@ -2,12 +2,6 @@
 
 #include <chrono>
 
-static std::chrono::time_point<std::chrono::system_clock> start_time = std::chrono::system_clock::now();
+std::chrono::time_point<std::chrono::system_clock> get_start_time();
 
-std::chrono::time_point<std::chrono::system_clock> get_start_time(){
-    return start_time;
-}
-
-void reset_start_time(){
-    start_time = std::chrono::system_clock::now();
-}
+void reset_start_time();
